+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - New features
-    - #81, #84: Support for specifying process-compose global CLI options
+    - #81, #84, #99: Support for specifying process-compose global CLI options
       - **Breaking changes**:
         - `preHook` and `postHook` are now inside `cli` module.
         - Old options `httpServer` and `tui` were removed; users should use the new `cli` module to set all process-compose cli global options. TUI can be disabled using `cli.environment.PC_DISABLE_TUI = true;`
@@ -11,6 +11,7 @@
     - #39: Allow `test` process to act as a test, which then gets run as part of flake checks.
     - #55: Add `lib` flake output - library of useful functions
         - #80: Add `evalModules`, to use process-compose-flake without flake-parts
+    - #102: Add `defaults.processSettings` option to set default settings for all processes
     - New options
       - #52: Add `is_foreground` option
       - ~~#54: Add `apiServer` option to control REST API server~~